backend/requirements.txt
django==5.1.4
python-dotenv
Pillow
djangorestframework
django-cors-headers
